work on parser

+++ includes/sphere.h
@@ -6,7 +6,7 @@
 /*   By: hpachy <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/09 14:17:41 by hpachy            #+#    #+#             */
-/*   Updated: 2017/03/16 13:36:05 by jrichard         ###   ########.fr       */
+/*   Updated: 2017/03/19 12:17:20 by jrichard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 typedef struct	s_sphere
 {
-	int			radius;
+	float		radius;
 	float		reflexion;
 }				t_sphere;
